% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenotypeLibrary-package.R
\docType{package}
\name{PhenotypeLibrarian-package}
\alias{PhenotypeLibrarian}
\alias{PhenotypeLibrarian-package}
\title{PhenotypeLibrarian: Librarian of Phenotypes}
\description{
Librarian of Phenotypes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/PhenotypeLibrarian}
  \item \url{https://github.com/OHDSI/PhenotypeLibrarian}
  \item Report bugs at \url{https://github.com/OHDSI/PhenotypeLibrarian/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
